split up main.ts
